Added support for mysql

+++ .entrypoint.sh
@@ -3,9 +3,14 @@
 echo "==============================================="
 chown -R www-data:www-data /var/www/html/storage /var/www/html/bootstrap/cache
 chmod -R 755 /var/www/html/storage /var/www/html/bootstrap/cache
-cp .env.example .env
+
+if [ ! -f .env ]; then
+  cp .env.example .env
+fi
+
 composer install
 
 php artisan key:generate
+php artisan migrate
 
 php-fpm
